move class related to Generating reports into their own package
