+++ src/main/java/org/tilkynna/report/generate/processengine/config/GenerateReportRetryPolicyConfig.java
@@ -4,7 +4,7 @@
  * License MIT: https://opensource.org/licenses/MIT
  * **************************************************
  */
-package org.tilkynna.report.generate.config;
+package org.tilkynna.report.generate.processengine.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
