+++ src/main/java/org/tilkynna/report/templates/ReportRequestThreadPoolConfig.java
@@ -4,7 +4,7 @@
  * License MIT: https://opensource.org/licenses/MIT
  * **************************************************
  */
-package org.tilkynna.report.generate.config;
+package org.tilkynna.report.templates;
 
 import java.util.concurrent.Executor;
 
@@ -24,7 +24,7 @@ public class ReportRequestThreadPoolConfig {
         ThreadPoolTaskExecutor threadPoolTaskScheduler = new ThreadPoolTaskExecutor();
 
         threadPoolTaskScheduler.setThreadPriority(Thread.MAX_PRIORITY);
-        threadPoolTaskScheduler.setThreadNamePrefix("GenReportReq-");
+        threadPoolTaskScheduler.setThreadNamePrefix("GenReportRequest-");
         threadPoolTaskScheduler.initialize();
         return threadPoolTaskScheduler;
     }
